tests for dotnet

+++ tests/util/fileExtension.js
@@ -59,7 +59,7 @@ function createDirIfNotPresent(path){
     return false
 }
 
-function rmContentsOfDir(dirPath) {
+function rmContentsOfDir(dirPath,extension) {
     try { var files = fs.readdirSync(dirPath); }
     catch(e) {
         console.log("Error reading directory %s due to %s", dirPath, e);
@@ -68,6 +68,8 @@ function rmContentsOfDir(dirPath) {
     if (files.length > 0)
         for (var i = 0; i < files.length; i++) {
         var filePath = _path.join(dirPath , files[i]);
+        if(extension && !filePath.endsWith(extension))
+            continue
         if (fs.statSync(filePath).isFile())
             fs.unlinkSync(filePath);
         else
